OPT: add save map extent  and fix some bugs.

+++ cs-front/src/main/java/cn/cangling/cs/client/rpc/repo/Repo.java
@@ -8,7 +8,7 @@ public class Repo {
     public String name;
     public Double lng;
     public Double lat;
-    public Double zoom;
+    public int zoom;
     public Double size;
     public String url;
     public boolean pared;
